requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
av==16.1.0
azure-cognitiveservices-speech==1.47.0
azure-core==1.38.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
distro==1.9.0
docstring_parser==0.17.0
eval_type_backport==0.3.1
filelock==3.20.3
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.1.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.1
Jinja2==3.1.6
jiter==0.12.0
livekit==1.0.23
livekit-agents==1.3.11
livekit-api==1.1.0
livekit-blingfire==1.1.0
livekit-plugins-azure==1.3.11
livekit-plugins-groq==1.3.11
livekit-plugins-openai==1.3.11
livekit-plugins-silero==1.3.11
livekit-plugins-turn-detector==1.3.11
livekit-protocol==1.1.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
nest-asyncio==1.6.0
numpy==2.4.1
onnxruntime==1.23.1
openai==2.15.0
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
packaging==25.0
pillow==12.1.0
prometheus_client==0.24.1
propcache==0.4.1
protobuf==6.33.4
psutil==7.2.1
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
rich==14.2.0
safetensors==0.7.0
shellingham==1.5.4
sniffio==1.3.1
sounddevice==0.5.3
sympy==1.14.0
tokenizers==0.22.2
tqdm==4.67.1
transformers==4.57.1
typer==0.21.1
types-protobuf==6.32.1.20251210
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0
zipp==3.23.0
